requirements.txt
anyio==3.6.1
astropy==5.1
astroquery==0.4.6
azure-functions==1.11.2
beautifulsoup4==4.11.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
fastapi==0.82.0
fonttools==4.37.3
greenlet==1.1.3
h11==0.13.0
html5lib==1.1
idna==3.3
jaraco.classes==3.2.2
jplephem==2.17
keyring==23.9.3
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.0
more-itertools==8.14.0
numba==0.56.2
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
plotly==5.10.0
poliastro==0.17.0
psycopg2==2.9.3
pydantic==1.10.2
pyerfa==2.0.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pyvo==1.3
PyYAML==6.0
requests==2.28.1
scipy==1.9.1
sgp4==2.21
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
starlette==0.19.1
tenacity==8.1.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
webencodings==0.5.1
